[NCL-4958] merged process-managers with bpm module and made BRPushManager to use DTO instead old rest-entity
