--- dto/src/main/java/org/jboss/pnc/dto/internal/bpm/BPMNotification.java
@@ -1,46 +1,0 @@
-/**
- * JBoss, Home of Professional Open Source.
- * Copyright 2014-2019 Red Hat, Inc., and individual contributors
- * as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jboss.pnc.dto.internal.bpm;
-
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-
-import lombok.Data;
-
-/**
- *
- * @author Honza Brázdil &lt;jbrazdil@redhat.com&gt;
- */
-@JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.EXISTING_PROPERTY,
-        property = "eventType")
-@JsonSubTypes({
-    @JsonSubTypes.Type(BuildResult.class),
-    @JsonSubTypes.Type(MilestoneReleaseResult.class),
-    @JsonSubTypes.Type(ProcessProgressUpdate.class),
-    @JsonSubTypes.Type(RepositoryCreationError.class),
-    @JsonSubTypes.Type(RepositoryCreationSuccess.class)
-})
-@Data
-public abstract class BPMNotification {
-
-    private final String eventType;
-
-    protected BPMNotification(String eventType) {
-        this.eventType = eventType;
-    }
-}
