--- dto/src/main/java/org/jboss/pnc/dto/internal/bpm/BPMTask.java
@@ -1,59 +1,0 @@
-/**
- * JBoss, Home of Professional Open Source.
- * Copyright 2014-2019 Red Hat, Inc., and individual contributors
- * as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jboss.pnc.dto.internal.bpm;
-
-import java.util.List;
-
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
-
-import lombok.Builder;
-import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
-
-/**
- *
- * @author Honza Brázdil &lt;jbrazdil@redhat.com&gt;
- */
-@Data
-@Builder(builderClassName = "Builder")
-@JsonDeserialize(builder = BPMTask.Builder.class)
-public class BPMTask {
-
-    /**
-     * Id of the task within BpmManager.
-     */
-    private final Integer taskId;
-
-    /**
-     * Id of the process instance within BPM engine.
-     */
-    private final Long processInstanceId;
-
-    /**
-     * Name of the process executing this task.
-     */
-    private final String processName;
-
-    private final List<BPMNotification> events;
-
-    @JsonPOJOBuilder(withPrefix = "")
-    public static final class Builder {
-    }
-}
