--- dto/src/main/java/org/jboss/pnc/dto/internal/bpm/RepositoryCreationError.java
@@ -1,49 +1,0 @@
-/**
- * JBoss, Home of Professional Open Source.
- * Copyright 2014-2019 Red Hat, Inc., and individual contributors
- * as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jboss.pnc.dto.internal.bpm;
-
-import com.fasterxml.jackson.annotation.JsonTypeName;
-
-import lombok.Data;
-import lombok.Getter;
-
-/**
- *
- * @author Honza Brázdil &lt;jbrazdil@redhat.com&gt;
- */
-@Data
-@JsonTypeName(value = RepositoryCreationError.RC_CREATION_ERROR)
-public class RepositoryCreationError extends BPMNotification {
-    static final String RC_CREATION_ERROR = "RC_CREATION_ERROR";
-
-    public RepositoryCreationError(Integer repositoryId, Integer buildConfigurationId, String errorMessage) {
-        super(RC_CREATION_ERROR);
-        this.repositoryId = repositoryId;
-        this.buildConfigurationId = buildConfigurationId;
-        this.errorMessage = errorMessage;
-    }
-
-    @Getter
-    private final Integer repositoryId;
-
-    @Getter
-    private final Integer buildConfigurationId;
-
-    @Getter
-    private final String errorMessage;
-}
