+++ ear-package/pom.xml
@@ -115,10 +115,6 @@
     </dependency>
     <dependency>
       <groupId>org.jboss.pnc</groupId>
-      <artifactId>process-managers</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.jboss.pnc</groupId>
       <artifactId>web</artifactId>
       <type>war</type>
     </dependency>
@@ -272,12 +268,6 @@
             <jarModule>
               <groupId>org.jboss.pnc</groupId>
               <artifactId>spi</artifactId>
-              <includeInApplicationXml>true</includeInApplicationXml>
-              <bundleDir>/</bundleDir>
-            </jarModule>
-            <jarModule>
-              <groupId>org.jboss.pnc</groupId>
-              <artifactId>process-managers</artifactId>
               <includeInApplicationXml>true</includeInApplicationXml>
               <bundleDir>/</bundleDir>
             </jarModule>
