+++ facade/src/main/java/org/jboss/pnc/facade/impl/BrewPusherImpl.java
@@ -27,8 +27,8 @@ import org.jboss.pnc.dto.BuildPushResult;
 import org.jboss.pnc.dto.requests.BuildPushRequest;
 import org.jboss.pnc.facade.BrewPusher;
 import org.jboss.pnc.mapper.api.BuildPushResultMapper;
-import org.jboss.pnc.managers.BuildResultPushManager;
-import org.jboss.pnc.managers.Result;
+import org.jboss.pnc.bpm.causeway.BuildResultPushManager;
+import org.jboss.pnc.bpm.causeway.Result;
 import org.jboss.pnc.model.BuildRecord;
 import org.jboss.pnc.model.BuildRecordPushResult;
 import org.jboss.pnc.spi.coordinator.ProcessException;
