+++ pom.xml
@@ -92,7 +92,6 @@
     <module>integration-test</module>
     <module>pnc-mock</module> <!-- keep 'pnc' prefix as 'mock' is dy default ignored in some cases (intellij) -->
     <module>bpm</module>
-    <module>process-managers</module>
     <module>processor</module>
     <module>rest-client</module>
     <module>mapper</module>
@@ -400,12 +399,6 @@
       <dependency>
         <groupId>org.jboss.pnc</groupId>
         <artifactId>bpm</artifactId>
-        <version>${project.version}</version>
-      </dependency>
-
-      <dependency>
-        <groupId>org.jboss.pnc</groupId>
-        <artifactId>process-managers</artifactId>
         <version>${project.version}</version>
       </dependency>
