--- process-managers/src/test/resources/logback.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    JBoss, Home of Professional Open Source.
-    Copyright 2014-2019 Red Hat, Inc., and individual contributors
-    as indicated by the @author tags.
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<configuration>
-
-  <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-    <layout class="ch.qos.logback.classic.PatternLayout">
-      <Pattern>
-        %d{yyyy-MM-dd HH:mm:ss} [%thread] %-5level %logger{36} ctx:%mdc - %msg%n
-      </Pattern>
-    </layout>
-  </appender>
-
-  <logger name="org.jboss.pnc" level="debug" additivity="false">
-    <appender-ref ref="STDOUT" />
-  </logger>
-
-  <root level="error">
-    <appender-ref ref="STDOUT" />
-  </root>
-
-</configuration>
