+++ rest/src/main/java/org/jboss/pnc/rest/endpoint/BuildRecordPushEndpoint.java
@@ -25,8 +25,8 @@ import org.jboss.pnc.common.Configuration;
 import org.jboss.pnc.common.json.ConfigurationParseException;
 import org.jboss.pnc.common.json.JsonOutputConverterMapper;
 import org.jboss.pnc.common.util.StringUtils;
-import org.jboss.pnc.managers.BuildResultPushManager;
-import org.jboss.pnc.managers.Result;
+import org.jboss.pnc.bpm.causeway.BuildResultPushManager;
+import org.jboss.pnc.bpm.causeway.Result;
 import org.jboss.pnc.model.BuildConfigurationAudited;
 import org.jboss.pnc.model.BuildRecord;
 import org.jboss.pnc.model.BuildRecordPushResult;
@@ -36,7 +36,6 @@ import org.jboss.pnc.rest.restmodel.BuildConfigSetRecordPushRequestRest;
 import org.jboss.pnc.rest.restmodel.BuildRecordPushRequestRest;
 import org.jboss.pnc.rest.restmodel.BuildRecordPushResultRest;
 import org.jboss.pnc.rest.restmodel.response.ResultRest;
-import org.jboss.pnc.rest.restmodel.response.error.ErrorResponseRest;
 import org.jboss.pnc.rest.validation.exceptions.RestValidationException;
 import org.jboss.pnc.spi.coordinator.ProcessException;
 import org.jboss.pnc.spi.datastore.predicates.BuildRecordPredicates;
